Add existing code developed under CCA effort
Current code is able to combine DSS files, alter the A, B, C, and F
parts of the paths, and default to a standard list of paths useful for
a CalSim3 run.

A CLI is included with the ability to specify the output DSS, a custom
FV file. A simple configuration file is established to communicate which
dss files to blend, and how to alter the paths so there are no
collisions. Currently, only collisions due to re-naming paths within a
single dss are detected. Other collisions occur silently.

Basic protections are also put into place for the CLI, for example, the
blended dss file must be new.